Fixed uint problem

+++ core/Server.cs
@@ -270,9 +270,9 @@ public partial class Server : Node
     /// Spawns a Node on the server and relays to all connections
     /// </summary>
     /// <returns>Node that was spawned</returns>
-    public static Node Spawn(uint prefabID, Vector3 position, Basis basis, Vector3 scale, NetworkConnection owner = null)
+    public static Node Spawn(int prefabID, Vector3 position, Basis basis, Vector3 scale, NetworkConnection owner = null)
     {
-        Node spawnedObject = NetworkManager.manager.NetworkObjectPrefabs[(int)prefabID].Instantiate();
+        Node spawnedObject = NetworkManager.manager.NetworkObjectPrefabs[prefabID].Instantiate();
         NetworkedNode netNode;
   
         // Finds its networked node, it should be a child of this spawned object
@@ -312,7 +312,7 @@ public partial class Server : Node
         SpawnNodePacket packet = new()
         {
             NetID = netNode.NetID,
-            prefabID = prefabID,
+            prefabID = (uint)prefabID,
             position = [position.X, position.Y, position.Z],
             rotation = [quat.X, quat.Y, quat.Z, quat.W],
             scale = [scale.X, scale.Y, scale.Z],
@@ -388,7 +388,7 @@ public partial class Server : Node
             if (NetworkManager.manager.PlayerPrefabID != -1)
             {
                 // Instantiate the player prefab if not -1
-                connection.playerObject = Spawn((uint)NetworkManager.manager.PlayerPrefabID, Vector3.Zero, Basis.Identity, Vector3.One, connection);
+                connection.playerObject = Spawn(NetworkManager.manager.PlayerPrefabID, Vector3.Zero, Basis.Identity, Vector3.One, connection);
                 connection.playerObject.Name = " [Conn ID: " + connection.GetRemoteID() + "]";
             }
         }
